Introduce 'explorer_dataset.ipynb' and refine 'cleaning the data.py'
This commit adds a new Jupyter notebook, 'explorer_dataset.ipynb', which enables us to comprehensively explore and understand the dataset such as counting the number of paintings per painter and giving summary statistics of image dimensions. At the same time, it applies a minor refinement to 'cleaning the data.py' by removing an unnecessary extra line.